exynos4-is: Add csis parent clock setup
With this patch the driver will set "parent" clock as a parent
clock of "mux" clock. When the samsung clocks driver is reworked
to use new composite clock type, the "mux" clock can be removed.

"parent" clock should be set in relevant dtsi file and can be
overwritten in a board dts file. This way it is ensured the
SCLK_CSIS has correct parent clock set, and the parent clock
can be selected per each board if required.

Note from Davids Paskevics:

This was originally in a downstream tree.

The patch has been modified to apply to a recent kernel
and to remove a workaround for non-DT platforms,
as these have been removed from the kernel in the meantime.

Signed-off-by: Sylwester Nawrocki <s.nawrocki@samsung.com>
Signed-off-by: Kyungmin Park <kyungmin.park@samsung.com>
Signed-off-by: Davids Paskevics <davids.paskevics@gmail.com>