of: reserved_mem: Use unflatten_devicetree APIs to scan reserved memory nodes
The unflatten_devicetree APIs have been setup and are available to be
used by the time the fdt_init_reserved_mem() function is called.
Since the unflatten_devicetree APIs are a more efficient way of scanning
through the DT nodes, switch to using these APIs to facilitate the rest
of the reserved memory processing.

Signed-off-by: Oreoluwa Babatunde <quic_obabatun@quicinc.com>
Link: https://lore.kernel.org/r/20240528223650.619532-4-quic_obabatun@quicinc.com
[robh: Drop 2nd check on address/size cell sizes]
Signed-off-by: Rob Herring (Arm) <robh@kernel.org>