drm/panel: Add backlight enable/disable for S6E63J0X03
Without this, the panel is always on at least minimum brightness,
but never turns off fully.

I'm not sure whether this is the right way to do this, but it works.

The downside is that visual snow is rendered on the panel until
the next time a frame is updated, but that should happen right after
unlocking the device so it might not be an issue in practice.