ACPI: HED: Always initialize before evged
When the HED driver is built-in, the HED init is behind evged as both
drivers are in the same initcall level, so their initialization order
is based on the Makefile order in which evged is first.  This causes
RAS records to be unhandled after evged has been initialized already
and before HED is initialized.

If the number of such RAS records is greater than the APEI HEST error
source number, the HEST resources may be exhausted which then may affect
subsequent RAS error reporting.

To address this issue, change the initcall level of HED to subsys_init
and to ensure that the initialization ordering between HED and evged
will always be preserved, change the ACPI_HED Kconfig option from
tristate to bool, which effectively prevents building HED as a module.

Reviewed-by: Jonathan Cameron <Jonathan.Cameron@huawei.com>
Signed-off-by: Xiaofei Tan <tanxiaofei@huawei.com>
Link: https://patch.msgid.link/20250117022957.25227-1-tanxiaofei@huawei.com
[ rjw: Subject and changelog rewrite ]
Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>