net: ax88796c: Do not pass negative values to pskb_expand_head()
Due to various implicit casts and being of type size_t,
the values of variables h and t could end up being so large
that they'd turn negative when cast to int
within the pskb_expand_head() call.

This triggered an assertion inside it,
resulting in an OOPS. Relevant excerpt:

    kernel BUG at net/core/skbuff.c:2232!
    Internal error: Oops - BUG: 0 [#1] SMP ARM
    Modules linked in:
    Tainted: [W]=WARN
    Hardware name: Samsung Exynos (Flattened Device Tree)
    Workqueue: events ax88796c_work
    PC is at pskb_expand_head (net/core/skbuff.c:2232
    LR is at ax88796c_work (drivers/net/ethernet/asix/ax88796c_main.c:279
    < snip >
    Call trace:
    (pskb_expand_head) from ax88796c_work (drivers/net/ethernet/asix/ax88796c_main.c:279)
    (ax88796c_work) from process_one_work (./include/linux/atomic/atomic-arch-fallback.h:457)
    (process_one_work) from worker_thread (kernel/workqueue.c:3316)
    (worker_thread) from kthread (kernel/kthread.c:463)
    (kthread) from ret_from_fork (arch/arm/kernel/entry-common.S:138)
    < snip >

Fix this by perfoming the entire calculation on ints.

Also make the code a bit more readable
by splitting the calculation into 2 steps.

Fixes: a97c69ba4f30e ("net: ax88796c: ASIX AX88796C SPI Ethernet Adapter Driver")